requirements.txt
python-dotenv==1.0.0
aiogram==2.25.1
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.5.0